+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- `tt pack `: added TCM file packaging.
+- `tt aeon connect`: add connection from the cluster config.
+- `tt aeon connect`: add connection from the `app:insance_name`.
+
 ### Changed
 
 ### Fixed